Clean



See merge request !343

+++ libinendi/plugins/plotting-filters/log_minmax/PVPlottingFilterLogMinmax.cpp
@@ -16,8 +16,8 @@ static void
 compute_log_plotting(pvcop::db::array const& mapped, pvcop::db::array const& minmax, uint32_t* dest)
 {
 	auto& mm = minmax.to_core_array<T>();
-	double ymin = mm[0];
-	double ymax = mm[1];
+	double ymin = (double)mm[0];
+	double ymax = (double)mm[1];
 
 	if (ymin == ymax) {
 		std::fill_n(dest, mapped.size(), 1UL << 31);
